requirements.txt

numpy
pandas
h5py
pytables
